+++ ChangeLog.md
@@ -8,6 +8,11 @@
   categories, external links, word counts, lifespan, streaks) with commas
   when they run into the thousands.
   ([#586](https://github.com/davep/blogmore/pull/586))
+- Added a responsive table of contents display to posts and pages, which
+  floats to the right on wider displays, or as a collapsible inline list on
+  narrow displays. Also added support for `show_toc` and `show_toc_inline`
+  frontmatter properties (on by default).
+  ([#588](https://github.com/davep/blogmore/pull/588))
 
 ## v2.37.0